+++ api-dist.js
@@ -1,5 +1,7 @@
 const Devour = require('devour-client');
 
+let apiAddressBattleBroadcast = '';
+let apiAddressTotalRanking = '';
 
 // To configure the external API from which the chatbot fetch keyworded responses
 // Change `apiaddress` to your external api address.
@@ -16,7 +18,20 @@ jsonApi.define('keyworded_response', {
     keyword: '',
     response: '',
 }, {
-    collectionPath: 'keyworded_response'
+    collectionPath: 'keyworded_responses'
 });
 
-module.exports = jsonApi;
+jsonApi.define('ignore_number', {
+    number: '',
+    note: '',
+}, {
+    collectionPath: 'ignore_numbers',
+});
+
+jsonApi.define('new_member_notice', {
+    response: '',
+}, {
+    collectionPath: 'new_member_notices',
+});
+
+module.exports = { jsonApi, apiAddressBattleBroadcast, apiAddressTotalRanking};