+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] 2019-08-31
+### Added
+- Added feature to ignore certain QQ number for keyword monitoring
+- Added feature for sending new member notices, fetching real time rankings, pk broadcast and etc.
+- Log errors to sentry 
+  - Added `@sentry/node@5.6.2` as a dependency
+
 ## [0.1.0] 2019-03-12
 ### Added
 - Added `README.md`